+++ src/design/plone/contenttypes/upgrades/upgrades.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from Acquisition import aq_base
 from collective.volto.blocksfield.field import BlocksField
 from copy import deepcopy
 from design.plone.contenttypes.upgrades.draftjs_converter import to_draftjs
@@ -473,18 +474,97 @@ def to_3101(context):
     fixed_total = 0
     for brain in api.content.find(portal_type="Documento"):
         item = brain.getObject()
-        for rel in getattr(item, 'servizi_collegati', []):
+        for rel in getattr(item, "servizi_collegati", []):
             service = rel.to_object
             if service:
-                service.altri_documenti.append(RelationValue(intids.getId(item)))
+                service.altri_documenti.append(
+                    RelationValue(intids.getId(item))
+                )
                 notify(ObjectModifiedEvent(service))
-                logger.info("Fixed item {}".format("/".join(service.getPhysicalPath())))
+                logger.info(
+                    "Fixed item {}".format("/".join(service.getPhysicalPath()))
+                )
 
-        if getattr(item, 'servizi_collegati', []):
+        if getattr(item, "servizi_collegati", []):
             delattr(item, "servizi_collegati")
             notify(ObjectModifiedEvent(item))
             fixed_total += 1
-            logger.info("Fixed item {}".format("/".join(item.getPhysicalPath())))
+            logger.info(
+                "Fixed item {}".format("/".join(item.getPhysicalPath()))
+            )
 
     logger.info("Fixing 'Documento': DONE")
     logger.info("Updated {} objects Documento".format(fixed_total))
+
+
+def to_volto13(context):  # noqa: C901
+    # convert listing blocks with new standard
+
+    logger.info("### START CONVERSION TO VOLTO 13 ###")
+
+    def fix_listing(blocks, url):
+        for block in blocks.values():
+            if block.get("@type", "") == "listing":
+                if block.get("template", False) and not block.get(
+                    "variation", False
+                ):
+                    # import pdb
+
+                    # pdb.set_trace()
+                    logger.error("- {}".format(url))
+                if block.get("template", False) and block.get(
+                    "variation", False
+                ):
+                    # import pdb
+
+                    # pdb.set_trace()
+                    logger.error("- {}".format(url))
+                if block.get("variation", "") == "default":
+                    block["variation"] = "simpleCard"
+                    logger.info("- {}".format(url))
+
+    # fix root
+    portal = api.portal.get()
+    portal_blocks = json.loads(portal.blocks)
+    fix_listing(portal_blocks, portal.absolute_url())
+    portal.blocks = json.dumps(portal_blocks)
+
+    # fix blocks in contents
+    pc = api.portal.get_tool(name="portal_catalog")
+    brains = pc()
+    tot = len(brains)
+    i = 0
+    for brain in brains:
+        i += 1
+        if i % 1000 == 0:
+            logger.info("Progress: {}/{}".format(i, tot))
+        item = aq_base(brain.getObject())
+        if getattr(item, "blocks", {}):
+            blocks = deepcopy(item.blocks)
+            if blocks:
+                fix_listing(blocks, brain.getURL())
+                item.blocks = blocks
+        for schema in iterSchemata(item):
+            # fix blocks in blocksfields
+            for name, field in getFields(schema).items():
+                if name == "blocks":
+                    blocks = deepcopy(item.blocks)
+                    if blocks:
+                        fix_listing(blocks, brain.getURL())
+                        item.blocks = blocks
+                elif isinstance(field, BlocksField):
+                    value = deepcopy(field.get(item))
+                    if not value:
+                        continue
+                    if isinstance(value, str):
+                        if value == "":
+                            setattr(
+                                item,
+                                name,
+                                {"blocks": {}, "blocks_layout": {"items": []}},
+                            )
+                            continue
+                    blocks = value.get("blocks", {})
+                    if blocks:
+                        fix_listing(blocks, brain.getURL())
+                        setattr(item, name, value)